+++ familyhelper-life-impl/src/main/java/com/dwarfeng/familyhelper/life/impl/util/FtpConstants.java
@@ -9,6 +9,7 @@ package com.dwarfeng.familyhelper.life.impl.util;
 public final class FtpConstants {
 
     public static final String PATH_PB_FILE = "familyhelper-life/pb-file";
+    public static final String PATH_ACTIVITY_TEMPLATE_COVER = "familyhelper-life/activity-template-cover";
 
     private FtpConstants() {
         throw new IllegalStateException("禁止实例化");