+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   - com.dwarfeng.familyhelper.life.stack.service.ActivityDataNodeOperateService。
   - com.dwarfeng.familyhelper.life.stack.service.ActivityDataItemOperateService。
   - com.dwarfeng.familyhelper.life.stack.service.ActivityTemplateOperateService。
+  - com.dwarfeng.familyhelper.life.stack.service.ActivityTemplateCoverOperateService。
 
 - 新增实体及其维护服务，单元测试通过。
   - com.dwarfeng.familyhelper.life.stack.bean.entity.Activity。